restructured bkk tee component for easier deploy
