+++ meta-bkk-display/recipes-app/bkk-tee/files/include/bkk_tee/bkk_tee_common_defs.h
