+++ meta-bkk-display/recipes-app/bkk-tee/files/src/trusted_app/CMakeLists.txt
