+++ meta-bkk-display/recipes-app/bkk-tee/files/src/trusted_app/sub.mk
@@ -1,0 +1,2 @@
+global-incdirs-y += ../../include/
+srcs-y += bkk_tee_ta.c
