+++ meta-bkk-display/recipes-app/bkk-tee/files/src/trusted_app/user_ta_header_defines.h
