--- meta-bkk-display/recipes-app/bkk-tee/files/trusted_app/sub.mk
@@ -1,2 +1,0 @@
-global-incdirs-y += ../common
-srcs-y += bkk_tee_ta.c
